+++ CHANGELOG.md
@@ -14,9 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `TaskSuite` base class and command line functionality for running [`checklist`](https://github.com/marcotcr/checklist) test suites, along with implementations for `SentimentAnalysisSuite`, `QuestionAnsweringSuite`, and `TextualEntailmentSuite`. These can be found in the `allennlp.sanity_checks.task_checklists` module.
-
-
-## Unreleased
+- Added `BiasMitigatorApplicator`, which wraps any Model and mitigates biases following training/finetuning
+on a downstream task.
 
 ### Fixed